requirements.txt
beautifulsoup4>=4.12.2
pandas>=2.1.0
selenium>=4.15.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
